+++ lang/en/theme_foundation.php
@@ -159,22 +159,8 @@ $string['loginbackgroundopacitydesc'] = 'Login background opacity for the login 
 // Syntax highlighter.
 $string['syntaxhighlightheading'] = 'Syntax highlighting settings';
 $string['syntaxhighlightheadingdesc'] = 'Set the syntax highlighting settings.';
-$string['syntaxhighlightdeprecatedwarning'] = 'Deprecation warning:';
-$string['syntaxhighlightdeprecatedinformation'] = 'Syntax highlighting in Foundation has been deprecated as it conflicts with my \'<a href="https://moodle.org/plugins/filter_synhi" target="_blank">SynHi</a>\' filter and will be removed in future releases.  Please use the \'SynHi\' filter instead and deactivate syntax highlighting here.';
-$string['syntaxhighlight'] = 'Activate syntax highlighting';
-$string['syntaxhighlightdesc'] = 'Activate syntax highlighting in courses.  When \'Yes\' you will be able to select the desired categories with the \'syntaxhighlightcat\' setting.';
-$string['syntaxhighlightpage'] = 'Syntax highlighting help';
-$string['syntaxhelpone'] = 'When editing anything (such as a label) with the text editor surround your code with a \'pre\' tag and add the class="brush: alias" where \'alias\' is one of the following:';
-$string['syntaxhelptwo'] = 'Brush name';
-$string['syntaxhelpthree'] = 'Brush alias';
-$string['syntaxhelpfour'] = 'For example:';
-$string['syntaxhelpfive'] = 'becomes:';
-$string['syntaxhelpsix'] = 'If you cannot see this help on a course then ask the administrator to activate \'Syntax highlighting\' for the courses\' category.';
-$string['syntaxhelpseven'] = 'Note: The \'&lt;\' and \'&gt;\' symbols seem to be problematic within in a \'pre\' tag, change them to \'&amp;lt;\' and \'&amp;gt;\' respectively for your code.  More information on';
-$string['syntaxhighlightcat'] = 'Syntax highlighting course categories';
-$string['syntaxhighlightcatdesc'] = 'Syntax highlighting in courses within the selected categories.  A help button that brings up a popup will be added to the footer of courses within the selected categories for course editors when editing.  To select none, use the \'Ctrl\' key.';
-$string['syntaxhighlightversion'] = 'Syntax highlighting version';
-$string['syntaxhighlightversiondesc'] = 'Select which version you would like to use.  Choose between 3.0.83 and 4.0.1.';
+$string['syntaxhighlightremoved'] = 'Syntax highlighting removed';
+$string['syntaxhighlightremoveddesc'] = 'Syntax highlighting functionality in Foundation has been removed.  Please use the \'<a href="https://moodle.org/plugins/filter_synhi" target="_blank">SynHi</a>\' filter instead.';
 
 // Carousels.
 $string['frontpagecarouselheading'] = 'Frontpage carousel';